requirements.txt
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.1.0
appdirs==1.4.4
appscript==1.3.0
asttokens==3.0.0
async-timeout==3.0.1
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloup==3.0.5
colour==0.1.5
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
dateparser==1.2.0
decorator==5.1.1
deprecation==2.1.0
diskcache==5.6.3
docx2pdf==0.1.8
et-xmlfile==1.1.0
executing==2.1.0
extra-streamlit-components==0.1.70
fonttools==4.55.3
frozendict==2.4.4
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.42
glcontext==3.0.0
gotrue==2.4.2
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.27.2
idna==3.6
importlib-metadata==7.0.1
ipython==8.31.0
isosurfaces==0.1.2
jedi==0.19.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.8
loadenv==0.1.1
lxml==5.2.1
macholib==1.16.3
manim==0.18.1
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
markdown2==2.4.13
Markdown2docx==0.1.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.0.3
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
networkx==3.4.2
numpy==1.26.4
openai==0.28.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pandoc==2.3
parso==0.8.4
patsy==0.5.6
peewee==3.17.3
pexpect==4.9.0
pillow==10.2.0
playwright==1.43.0
plotly==5.19.0
plotly-express==0.4.1
plumbum==1.8.2
ply==3.11
postgrest==0.16.4
prompt_toolkit==3.0.48
protobuf==4.25.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==15.0.0
pycairo==1.27.0
pycryptodome==3.19.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.1.0
pyglet==2.0.20
PyGLM==2.7.3
Pygments==2.17.2
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyJWT==2.8.0
pyobjc-core==10.3.2
pyobjc-framework-Cocoa==10.3.2
PyOpenGL==3.1.7
pyparsing==3.2.1
PyPDF2==3.0.1
pyperclip==1.9.0
PySocks==1.7.1
python-binance==1.0.19
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.1
PyYAML==6.0.1
realtime==1.0.4
referencing==0.33.0
regex==2023.10.3
reportlab==4.2.5
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.12.0
screeninfo==0.8.1
selenium==4.16.0
setuptools==68.2.2
six==1.16.0
skia-pathops==0.8.0.post2
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
srt==3.5.3
stack-data==0.6.3
statsmodels==0.14.1
storage3==0.7.4
streamlit==1.31.1
streamlit-authenticator==0.3.1
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
svgelements==1.9.6
sympy==1.13.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.8.0
urllib3==2.2.1
validators==0.22.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wheel==0.41.2
WooCommerce==3.0.0
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.38
zipp==3.17.0
